feat: integración completa con backend propio, gestión de favoritos local y mejoras UX
Refactoriza todas las llamadas de datos para pasar exclusivamente por el backend propio.
Añade gestión de favoritos en localStorage y sincronización con el estado global.
Mejora la experiencia de usuario en login/register y manejo de modales.
Corrige el filtrado y búsqueda de Pokémon usando endpoints centralizados.
Actualiza la documentación para reflejar la nueva arquitectura y dependencias.
Ajusta el preloader y el flujo de autenticación para mayor robustez.
Limpia reglas de ESLint y actualiza configuraciones para React 18.
