+++ src/components/App/App.css
@@ -1,12 +1,5 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-/* Estilos específicos del componente App */
 .app {
-  max-height: 100vh;
+  height: calc(100vh - 40px);
   display: flex;
   flex-direction: column;
   background-color: var(--color-background);
@@ -18,7 +11,7 @@
   display: grid;
   place-items: center;
   background-color: var(--color-background);
-  z-index: 3; /* Ajustado dentro del rango permitido */
+  z-index: 3;
 }
 
 .app__header {
