+++ src/components/Footer/Footer.css
@@ -3,7 +3,7 @@
   padding: 20px 0;
   margin-top: auto;
   position: relative;
-  z-index: 0; /* Asegurar que el footer no interfiera con el popup */
+  z-index: 0;
 }
 
 .footer__container {
