+++ src/components/Header/Header.css
@@ -1,15 +1,15 @@
 .header {
   background: linear-gradient(135deg, var(--color-accent) 0%, var(--color-accent-dark) 100%);
-  padding: 5px; /* Reducir aún más el padding para disminuir la altura */
+  padding: 5px;
   box-shadow: var(--shadow-sm);
-  position: relative; /* Cambié de sticky a relative para quitar el efecto sticky */
+  position: relative;
   top: 0;
   display: flex;
-  flex-direction: row; /* Cambiar a fila */
-  align-items: center; /* Alinear todos los elementos verticalmente */
-  justify-content: space-between; /* Espaciado entre los elementos */
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
   flex-wrap: wrap;
-  gap: 20px; /* Espaciado entre los elementos */
+  gap: 20px;
 }
 
 .header__container {
@@ -18,28 +18,28 @@
   padding: 0 20px;
   display: flex;
   flex: 1;
-  flex-direction: row; /* Cambiar a fila para alinear los elementos horizontalmente */
-  align-items: center; /* Asegurar que los elementos estén alineados verticalmente */
-  justify-content: space-between; /* Espaciado adecuado entre los elementos */
-  gap: 10px; /* Reducir aún más el espacio entre elementos */
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px;
 }
 
 .header__title {
   color: var(--color-text-light);
-  margin: 0; /* Eliminar márgenes adicionales */
+  margin: 0;
   font-size: clamp(1.5rem, 4vw, 2rem);
   font-weight: bold;
   text-shadow: 2px 2px 4px rgb(0 0 0 / 20%);
   white-space: nowrap;
   margin-bottom: 10px;
-  margin-right: auto; /* Alinear el título a la izquierda */
+  margin-right: auto;
 }
 
 .header__filters {
   display: flex;
   gap: 10px;
   margin-top: 10px;
-  align-items: center; /* Asegurar que los filtros estén alineados verticalmente */
+  align-items: center;
 }
 
 .header__filter {
@@ -63,8 +63,8 @@
 }
 
 .navigation {
-  margin-left: auto; /* Alinear la navegación a la derecha */
-  align-self: center; /* Asegurar que la navegación esté alineada verticalmente */
+  margin-left: auto;
+  align-self: center;
 }
 
 @media (max-width: 768px) {
