+++ src/components/Main/Main.css
@@ -1,16 +1,14 @@
-/* Estilos específicos del componente Main */
 .main {
   position: relative;
-  height: 500px;
+  height: calc(100vh - 200px);
   background-color: var(--color-background);
+  overflow: auto;
 }
 
 .main__content {
   max-width: 1280px;
-  height: 400px;
   margin: 0 auto;
   padding: 40px 20px;
-  overflow: auto;
 }
 
 .main__preloader {
@@ -24,12 +22,12 @@
 .main__preloader-overlay {
   position: absolute;
   inset: 0;
-  z-index: 2000;
-  background: #fff !important;
+  z-index: 3;
+  background: #fff;
   display: flex;
   align-items: center;
   justify-content: center;
-  opacity: 1 !important;
+  opacity: 1;
 }
 
 @media (max-width: 768px) {
