+++ src/components/PokemonList/PokemonList.css
@@ -1,4 +1,3 @@
-/* Estilos específicos del componente PokemonList */
 .pokemon-container {
   padding: 20px;
   position: relative;
@@ -19,7 +18,7 @@
   inset: 0;
   display: grid;
   place-items: center;
-  z-index: 1; /* Mantener el valor permitido */
+  z-index: 1;
 }
 
 .pokemon-loading {
