+++ src/components/PopupWithForm/PopupWithForm.css
@@ -1,52 +1,29 @@
 .popup-overlay {
-  position: fixed;
+  position: absolute;
   inset: 0;
-  background-color: #000 !important;
-  opacity: 1 !important; /* Asegurar que no haya transparencia */
+  opacity: 1;
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 2; /* Mantener el valor permitido */
-  pointer-events: auto; /* Permitir interacción solo con el popup */
-}
-
-.popup-overlay .popup-loading {
-  pointer-events: all; /* Permitir interacción con el preloader */
-}
-
-body.popup-open {
-  overflow: hidden;
-}
-
-.hidden-background {
-  pointer-events: none; /* Deshabilitar interacción con el fondo */
-  filter: none; /* Mantener el fondo visible sin cambios */
-  opacity: 1; /* Asegurar que las imágenes permanezcan visibles */
-}
-
-/* Asegurar que el contenido principal no desaparezca */
-#root {
-  visibility: visible; /* Mantener visible el contenido principal */
+  max-width: 1280px;
+  max-height: 500px;
+  z-index: 3;
+  pointer-events: auto;
 }
 
 .popup-content {
-  background-color: #ffffff !important; /* Fondo blanco sólido */
+  background-color: #ffffff;
   border-radius: 20px;
-  padding: 10px; /* Reducir el padding para optimizar el espacio */
+  padding: 10px;
   position: relative;
-  width: 90%;
-  max-width: 600px;
-  max-height: 85vh; /* Ajusté la altura máxima para evitar cortes */
-  overflow: hidden; /* Evitar el scroll */
+  width: 100%;
+  max-width: 1000px;
+  max-height: 85vh;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  box-shadow: 0 4px 20px 0 rgba(0, 0, 0, 1) !important; /* Sombra completamente opaca */
-  opacity: 1 !important; /* Sin transparencia */
-}
-
-.popup-content * {
-  opacity: 1 !important; /* Forzar opacidad de todos los elementos secundarios */
+  box-shadow: 0 4px 20px 0 rgba(0, 0, 0, 1);
+  opacity: 1;
 }
 
 .popup-close {
@@ -68,7 +45,7 @@ body.popup-open {
   justify-content: center;
   align-items: center;
   height: 100%;
-  opacity: 0.8; /* Transparencia permitida */
+  opacity: 0.8;
 }
 
 .pokeball-animation {
@@ -79,4 +56,156 @@ body.popup-open {
   height: 100%;
   background: radial-gradient(circle, #f00 50%, #fff 50%);
   border-radius: 50%;
+}
+
+.pokemon-details {
+  display: flex;
+  align-items: center;
+  padding: 20px 0 10px 0;
+}
+
+.pokemon-header {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 10px;
+}
+
+.pokemon-header h2 {
+  font-size: 2rem;
+  margin: 0;
+  text-transform: capitalize;
+}
+
+.pokemon-header span {
+  font-size: 1.2rem;
+  color: #888;
+}
+
+.popup-image-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 180px;
+  height: 180px;
+  margin: 0 auto 10px auto;
+}
+
+.pokemon-detail-image {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+}
+
+.pokemon-info {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 18px;
+}
+
+.pokemon-types {
+  display: flex;
+  gap: 8px;
+  justify-content: center;
+  margin-bottom: 8px;
+}
+
+.type-badge {
+  padding: 4px 12px;
+  border-radius: 12px;
+  color: #fff;
+  font-size: 0.95rem;
+  text-transform: capitalize;
+  font-weight: 600;
+  background: #aaa;
+  box-shadow: 0 1px 4px rgba(0,0,0,0.08);
+}
+/* Ejemplo de colores para algunos tipos */
+.type-badge.fire { background: #ee8130; }
+.type-badge.water { background: #6390f0; }
+.type-badge.grass { background: #7ac74c; }
+.type-badge.electric { background: #f7d02c; color: #333; }
+.type-badge.ice { background: #96d9d6; color: #333; }
+.type-badge.fighting { background: #c22e28; }
+.type-badge.poison { background: #a33ea1; }
+.type-badge.ground { background: #e2bf65; color: #333; }
+.type-badge.flying { background: #a98ff3; }
+.type-badge.psychic { background: #f95587; }
+.type-badge.bug { background: #a6b91a; }
+.type-badge.rock { background: #b6a136; color: #333; }
+.type-badge.ghost { background: #735797; }
+.type-badge.dragon { background: #6f35fc; }
+.type-badge.dark { background: #705746; }
+.type-badge.steel { background: #b7b7ce; color: #333; }
+.type-badge.fairy { background: #d685ad; }
+
+.pokemon-stats {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 8px;
+}
+
+.stat-row {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.stat-name {
+  width: 80px;
+  text-transform: capitalize;
+  font-size: 1rem;
+  color: #444;
+}
+
+.stat-bar-container {
+  flex: 1;
+  background: #eee;
+  border-radius: 8px;
+  height: 16px;
+  overflow: hidden;
+  margin: 0 6px;
+}
+
+.stat-bar {
+  height: 100%;
+  border-radius: 8px;
+  background: linear-gradient(90deg, #4caf50 60%, #81c784 100%);
+  transition: width 0.5s cubic-bezier(0.4,0,0.2,1);
+}
+
+.stat-value {
+  width: 32px;
+  text-align: right;
+  font-size: 1rem;
+  color: #222;
+}
+
+.pokemon-abilities {
+  margin-top: 8px;
+}
+
+.pokemon-abilities h3 {
+  font-size: 1.1rem;
+  margin: 0 0 4px 0;
+  color: #555;
+}
+
+.abilities-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+}
+
+.ability-badge {
+  background: #f1c40f;
+  color: #222;
+  border-radius: 10px;
+  padding: 3px 10px;
+  font-size: 0.95rem;
+  font-weight: 500;
+  text-transform: capitalize;
+  box-shadow: 0 1px 2px rgba(0,0,0,0.07);
 }
