+++ src/index.css
@@ -1,6 +1,4 @@
-/* Estilos globales que afectan a toda la página */
 :root {
-  /* Variables globales */
   --color-text-light: #ffffff;
   --color-text-dark: #000000;
   --color-primary: #213547;
@@ -13,7 +11,15 @@
   --font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
 }
 
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  height: calc(100vh - 64px);
+}
+
 body {
+  height: 100vh;
   margin: 0;
   font-family: var(--font-family);
   background-color: var(--color-background);
